Remove extra task for assertion layer
