+++ plugin/src/main/kotlin/com/jraska/module/graph/assertion/Api.kt
@@ -6,9 +6,8 @@ object Api {
 
     const val ASSERT_ALL = "assertModulesGraph"
     const val ASSERT_MAX_HEIGHT = "assertMaxHeight"
-    const val ASSERT_LAYER_ORDER = "assertModuleLayersOrder"
-    const val ASSERT_NO_IN_LAYER_PREFIX = "assertNoDependenciesWithin"
-    const val ASSERT_USER_RULES = "assertRestricted"
+    const val ASSERT_LAYER_ORDER = "assertModuleLayers"
+    const val ASSERT_USER_RULES = "assertRestrictions"
   }
 
   object Parameters {
